scripts/requirements.txt
cadquery==2.4.0
pyocp>=7.7.2.1
